+++ src/mcp_evals/mcp_client/__init__.py
@@ -1,5 +1,12 @@
 from mcp_evals.mcp_client.discover import discover_tools, load_server_config
 from mcp_evals.mcp_client.http import discover_http
+from mcp_evals.mcp_client.sse import discover_sse
 from mcp_evals.mcp_client.stdio import discover_stdio
 
-__all__ = ["discover_http", "discover_stdio", "discover_tools", "load_server_config"]
+__all__ = [
+    "discover_http",
+    "discover_sse",
+    "discover_stdio",
+    "discover_tools",
+    "load_server_config",
+]